ADD join alerts
