+++ main.py
@@ -5,7 +5,7 @@ import logging
 import Config as cfg
 from cmds.admingroup import Admin
 from cmds.utils import Utils
-from filters.welcomeMessage import WelcomeMessage
+from filters.newMember import NewMember
 from cmds.maingroup import Maingroup
 from filters.adminAlert import Admin_Alert, AdminAlert
 
@@ -20,13 +20,13 @@ print("Running")
 dispatcher.add_handler(CommandHandler('arules', Admin.aRules))
 dispatcher.add_handler(CommandHandler('id', Utils.id))
 dispatcher.add_handler(CallbackQueryHandler(Admin.button))
-dispatcher.add_handler(MessageHandler(Filters.status_update.new_chat_members, WelcomeMessage.greeting))
 dispatcher.add_handler(ConversationHandler(
     entry_points=[CommandHandler('init', Admin.init)],
     states={1: [RegexHandler(r'-?\d+', Admin.link)]},
     fallbacks=[]
 ))
 
+dispatcher.add_handler(MessageHandler(Filters.status_update.new_chat_members, NewMember.joinAlert))
 dispatcher.add_handler(MessageHandler(Admin_Alert, AdminAlert.admAlert))
 dispatcher.add_handler(CommandHandler('rules', Maingroup.rules))
